Move ProcessMessageOptions types into its own file
