+++ packages/core/characterisation-tests/helpers/processMessage.ts
@@ -2,21 +2,10 @@ import type Phase1Result from "../../phase1/types/Phase1Result"
 import type Phase2Result from "../../phase2/types/Phase2Result"
 import type Phase3Result from "../../phase3/types/Phase3Result"
 import type PncUpdateRequestError from "../../phase3/types/PncUpdateRequestError"
-import type { ResultedCaseMessageParsedXml } from "../../types/SpiResult"
+import type { ProcessMessageOptions } from "../types/ProcessMessageOptions"
 
 import Phase from "../../types/Phase"
 import { processMessageCorePhase1, processMessageCorePhase2, processMessageCorePhase3 } from "./processMessageCore"
-
-export type ProcessMessageOptions = {
-  expectRecord?: boolean
-  phase?: Phase
-  pncAdjudication?: boolean
-  pncCaseType?: string
-  pncErrorMessage?: string
-  pncMessage?: string
-  pncOverrides?: Partial<ResultedCaseMessageParsedXml>
-  recordable?: boolean
-}
 
 export const processPhase1Message = (messageXml: string, options: ProcessMessageOptions = {}): Promise<Phase1Result> =>
   processMessageCorePhase1(messageXml, { phase: Phase.HEARING_OUTCOME, ...options })
